+++ aws-java-sdk-inspector/src/main/java/com/amazonaws/services/inspector/model/AssessmentRunState.java
@@ -31,7 +31,8 @@ public enum AssessmentRunState {
     FAILED("FAILED"),
     ERROR("ERROR"),
     COMPLETED("COMPLETED"),
-    COMPLETED_WITH_ERRORS("COMPLETED_WITH_ERRORS");
+    COMPLETED_WITH_ERRORS("COMPLETED_WITH_ERRORS"),
+    CANCELED("CANCELED");
 
     private String value;