+++ aws-java-sdk-ssm/src/main/java/com/amazonaws/services/simplesystemsmanagement/AWSSimpleSystemsManagement.java
@@ -548,7 +548,7 @@ public interface AWSSimpleSystemsManagement {
 
     /**
      * <p>
-     * Delete a list of parameters.
+     * Delete a list of parameters. This API is used to delete parameters by using the Amazon EC2 console.
      * </p>
      * 
      * @param deleteParametersRequest
@@ -1124,6 +1124,13 @@ public interface AWSSimpleSystemsManagement {
      * <p>
      * Get information about a parameter.
      * </p>
+     * <p>
+     * Request results are returned on a best-effort basis. If you specify <code>MaxResults</code> in the request, the
+     * response includes information up to the limit specified. The number of items returned, however, can be between
+     * zero and the value of <code>MaxResults</code>. If the service reaches an internal limit while processing the
+     * results, it stops the operation and returns the matching values up to that point and a <code>NextToken</code>.
+     * You can specify the <code>NextToken</code> in a subsequent call to get the next set of results.
+     * </p>
      * 
      * @param describeParametersRequest
      * @return Result of the DescribeParameters operation returned by the service.
@@ -1456,6 +1463,13 @@ public interface AWSSimpleSystemsManagement {
      * Retrieve parameters in a specific hierarchy. For more information, see <a
      * href="http://docs.aws.amazon.com/systems-manager/latest/userguide/sysman-paramstore-working.html">Working with
      * Systems Manager Parameters</a>.
+     * </p>
+     * <p>
+     * Request results are returned on a best-effort basis. If you specify <code>MaxResults</code> in the request, the
+     * response includes information up to the limit specified. The number of items returned, however, can be between
+     * zero and the value of <code>MaxResults</code>. If the service reaches an internal limit while processing the
+     * results, it stops the operation and returns the matching values up to that point and a <code>NextToken</code>.
+     * You can specify the <code>NextToken</code> in a subsequent call to get the next set of results.
      * </p>
      * 
      * @param getParametersByPathRequest
@@ -1959,6 +1973,25 @@ public interface AWSSimpleSystemsManagement {
      *      Documentation</a>
      */
     RemoveTagsFromResourceResult removeTagsFromResource(RemoveTagsFromResourceRequest removeTagsFromResourceRequest);
+
+    /**
+     * <p>
+     * Sends a signal to an Automation execution to change the current behavior or status of the execution.
+     * </p>
+     * 
+     * @param sendAutomationSignalRequest
+     * @return Result of the SendAutomationSignal operation returned by the service.
+     * @throws AutomationExecutionNotFoundException
+     *         There is no automation execution information for the requested automation execution ID.
+     * @throws InvalidAutomationSignalException
+     *         The signal is not valid for the current Automation execution.
+     * @throws InternalServerErrorException
+     *         An error occurred on the server side.
+     * @sample AWSSimpleSystemsManagement.SendAutomationSignal
+     * @see <a href="http://docs.aws.amazon.com/goto/WebAPI/ssm-2014-11-06/SendAutomationSignal" target="_top">AWS API
+     *      Documentation</a>
+     */
+    SendAutomationSignalResult sendAutomationSignal(SendAutomationSignalRequest sendAutomationSignalRequest);
 
     /**
      * <p>